ultradns_record: extract common, cleanup
